requirements.txt
aiogram==3.10.0
pandas==2.2.2
numpy==1.26.4
xlsxwriter==3.2.0
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1